Feat/payment

+++ database/migrations/2023_08_21_014013_create_orders_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->unsignedBigInteger('product_id');
             $table->unsignedBigInteger('addresses_id');
             $table->integer('qty');
-            $table->decimal('total_price', 10, 2);
+            $table->double('total_price', 10, 2);
             $table->string('status');
             $table->timestamps();
